feat(6922): extract phase6 archival modules
