+++ crates/kamn-core/src/data_layer_m10_partition_archival/phase6/adapters.rs
@@ -1,0 +1,2 @@
+pub(super) mod bridge;
+pub(super) mod error_mapping;
